third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/last-execution-context-expected.txt

Tests how execution context and target are selected.


Adding page target
Execution context selected: contentScript1
Target selected: mock-target-page-target
Execution context selected: mock-target-page-target:42

Adding sw target

Removing page main frame
Execution context selected: contentScript1

Readding page main frame
Execution context selected: mock-target-page-target:42

Switching to sw target
Target selected: mock-target-sw-target
Execution context selected: mock-target-sw-target:

Switching to page target
Target selected: mock-target-page-target
Execution context selected: mock-target-page-target:42

User selected content script
Execution context selected: contentScript1

Switching to sw target
Target selected: mock-target-sw-target
Execution context selected: mock-target-sw-target:

Switching to page target
Target selected: mock-target-page-target
Execution context selected: mock-target-page-target:42

User selected iframe1
Execution context selected: mock-target-page-target:iframe1

Switching to sw target
Target selected: mock-target-sw-target
Execution context selected: mock-target-sw-target:

Switching to page target
Target selected: mock-target-page-target
Execution context selected: mock-target-page-target:iframe1

Switching to sw target
Target selected: mock-target-sw-target
Execution context selected: mock-target-sw-target:

Removing page main frame

Readding page main frame
